128要素でもＯＫ、７階層でもＯＫ，初期化の width を length にしてしまっていた

+++ Assets/NnJob.cs
@@ -15,7 +15,7 @@ using System.Numerics;
 namespace nn
 {
 
-    using number = System.Double;
+    using number = System.Single;
     //using Unity.VisualScripting;
